requirements.txt
Django
gunicorn
django-environ
selenium
whitenoise
pytest-django
djlint
black
django-sesame
flake8
